+++ packages/terra-card/src/Card.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
-import styles from './Card.scss';
+import styles from './Card.module.scss';
 import CardBody from './CardBody';
 
 const cx = classNames.bind(styles);