+++ packages/terra-embedded-content-consumer/src/terra-dev-site/test/embedded-content-consumer/providers/DataStatusProvider.test.jsx
@@ -2,6 +2,7 @@
 import React from 'react';
 import { Provider } from 'xfc';
 import './Provider.scss';
+import './ProviderIframe.module.scss';
 
 class DataStatusProvider extends React.Component {
   constructor(props) {