+++ packages/terra-form-textarea/src/Textarea.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 
-import styles from './Textarea.scss';
+import styles from './Textarea.module.scss';
 
 const cx = classNames.bind(styles);