+++ packages/terra-section-header/src/SectionHeader.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 import Arrange from 'terra-arrange';
-import styles from './SectionHeader.scss';
+import styles from './SectionHeader.module.scss';
 
 const cx = classNames.bind(styles);