+++ packages/terra-grid/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.16.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 4.15.0 - (July 17, 2018)
 ------------------
 ### Changed