+++ packages/terra-section-header/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Added
 * Added isTransparent prop to set background to be transparent
 
+1.16.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 1.15.0 - (July 17, 2018)
 ------------------
 ### Changed