+++ packages/terra-signature/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+1.20.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 1.19.0 - (July 17, 2018)
 ------------------
 ### Changed