+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+3.13.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+
 3.12.0 - (July 17, 2018)
 ------------------
 ### Changed