+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog
 Unreleased
 ----------
 
+2.19.0 - (July 19, 2018)
+------------------
+### Changed
+* Add .module extension to SCSS file(s)
+* Replace legacy Input with the new terra-form-input package
+* Give desktop time input the same styles as terra-form-input.
+
 2.18.0 - (July 17, 2018)
 ------------------
 ### Changed